refactor: 게임 시작 함수 리팩토링
