+++ src/main/java/baseball/BaseballException.java
@@ -1,4 +1,6 @@
 package baseball;
 
 public class BaseballException {
+
+
 }
